Reformatted about page into a component
