+++ src/app/about/page.js
@@ -1,17 +1,11 @@
 import "@/app/layout.js";
+import About from "@/components/about.jsx";
 
-const About = () => {
+const AboutPage = () => {
   return (
-    <div className="bg-purplecustom min-h-screen flex items-center justify-center">
-      <div className="flex flex-col tracking-widest">
-        <div className="mb-1 text-9xl text-center text-purple-200">
-          <p>ABOUT</p>
-        </div>
-        <div className="mb-1 text-9xl text-center text-purple-200">
-          <p>US</p>
-        </div>
-      </div>
+    <div>
+      <About />
     </div>
   );
 };
-export default About;
+export default AboutPage;
