+++ src/components/about.jsx
@@ -1,0 +1,15 @@
+const About = () => {
+  return (
+    <div className="bg-purplecustom min-h-screen flex items-center justify-center">
+      <div className="flex flex-col tracking-widest">
+        <div className="mb-1 text-9xl text-center text-purple-200">
+          <p>ABOUT</p>
+        </div>
+        <div className="mb-1 text-9xl text-center text-purple-200">
+          <p>US</p>
+        </div>
+      </div>
+    </div>
+  );
+};
+export default About;
